Spring practice_hw_10. Add BigDecimal instead of int.

+++ api/src/main/java/com/naumovets/market/api/dto/cart/CartDto.java
@@ -1,10 +1,12 @@
 package com.naumovets.market.api.dto.cart;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class CartDto {
     private List<CartItemDto> list;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 }
