+++ api/src/main/java/com/naumovets/market/api/dto/cart/CartItemDto.java
@@ -2,11 +2,13 @@ package com.naumovets.market.api.dto.cart;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CartItemDto {
     private Long productId;
     private String productTitle;
     private int quantity;
-    private int pricePerProduct;
-    private int price;
+    private BigDecimal pricePerProduct;
+    private BigDecimal price;
 }
