+++ api/src/main/java/com/naumovets/market/api/dto/product/ProductDto.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductDto {
     private Long id;
     private String title;
-    private int cost;
+    private BigDecimal cost;
     private String categoryTitle;
 }
